Refactoring pulse_trigger, moving two classes into pulse_modules
